alreadyMinified

+++ gulpfile.js
@@ -93,7 +93,8 @@ var paths = {
         icons: "src/img/icons",
         firebase: "src/lib/firebase/**/*",
         js: "src/js/*.js",
-        serviceWorker: "src/firebase-messaging-sw.js"
+        serviceWorker: "src/firebase-messaging-sw.js",
+        staticData: 'src/data/qmStaticData.js'
     },
     www: {
         devCredentials: "www/dev-credentials.json",
@@ -102,7 +103,8 @@ var paths = {
         defaultPrivateConfig: "www/default.private_config.json",
         icons: "www/img/icons",
         firebase: "www/lib/firebase/",
-        js: "www/js/"
+        js: "www/js/",
+        staticData: 'src/data/qmStaticData.js'
     },
     chcpLogin: '.chcplogin'
 };
@@ -447,10 +449,27 @@ var qmGulp = {
     },
     buildInfoHelper: {
         alreadyMinified: function(){
-            if(!qmGulp.buildInfoHelper.getPreviousBuildInfo().gitCommitShaHash){return false;}
-            return qmGulp.buildInfoHelper.getCurrentBuildInfo().gitCommitShaHash === qmGulp.buildInfoHelper.getCurrentBuildInfo().gitCommitShaHash;
+            var previousSha = qmGulp.buildInfoHelper.getPreviousBuildSha();
+            if(!previousSha){
+                qmLog.error("Could not get previous git commit SHA!");
+                return false;
+            }
+            var currentSha = qmGit.getCurrentGitCommitSha();
+            if(!currentSha){
+                qmLog.error("Could not get current git commit SHA!");
+                return false;
+            }
+            var alreadyMinified = previousSha === currentSha;
+            if(!alreadyMinified){
+                qmLog.info("current sha " + currentSha + " and previous commit SHA " + previousSha +
+                    " don't match so we need to minify again");
+            } else {
+                qmLog.info("No need to minify again because current sha " + currentSha + " and previous commit SHA " +
+                    previousSha + " match");
+            }
+            return alreadyMinified;
         },
-        previousBuildInfo: {
+        buildInfo: {
             iosCFBundleVersion: null,
             builtAt: null,
             buildServer: null,
@@ -459,6 +478,11 @@ var qmGulp = {
             versionNumbers: null,
             gitBranch: null,
             gitCommitShaHash: null
+        },
+        getPreviousBuildSha: function(){
+            var previousBuildInfo = qmGulp.buildInfoHelper.getPreviousBuildInfo();
+            if(!previousBuildInfo){return false;}
+            return previousBuildInfo.gitCommitShaHash;
         },
         getCurrentBuildInfo: function () {
             return qmGulp.buildInfoHelper.currentBuildInfo = {
@@ -473,8 +497,14 @@ var qmGulp = {
             };
         },
         getPreviousBuildInfo: function () {
-            return JSON.parse(fs.readFileSync(paths.www.buildInfo));
+            var previousBuildInfo = readFile(paths.src.buildInfo);
+            if(!previousBuildInfo){
+                qmLog.info("No previous BuildInfo file at "+paths.src.buildInfo);
+                return qmGulp.buildInfoHelper.previousBuildInfo = false;
+            }
+            return qmGulp.buildInfoHelper.previousBuildInfo = previousBuildInfo;
         },
+        previousBuildInfo: null,
         writeBuildInfo: function () {
             var buildInfo = qmGulp.buildInfoHelper.getCurrentBuildInfo();
             writeToFile(paths.src.buildInfo, buildInfo);
@@ -658,7 +688,12 @@ function readDevCredentials(){
     }
 }
 function readFile(path){
-    return JSON.parse(fs.readFileSync(path));
+    try {
+        return JSON.parse(fs.readFileSync(path));
+    } catch (e) {
+        qmLog.error("Could not read "+path);
+        return false;
+    }
 }
 function outputFileContents(path){
     qmLog.info(path+": "+fs.readFileSync(path))
@@ -1275,7 +1310,10 @@ function writeToFileWithCallback(filePath, stringContents, callback) {
     if(typeof stringContents !== "string"){stringContents = JSON.stringify(stringContents);}
     return fs.writeFile(filePath, stringContents, callback);
 }
-gulp.task('createSuccessFile', function () {return fs.writeFileSync('success');});
+gulp.task('createSuccessFile', function () {
+    writeToFile('lastCommitBuilt', qmGit.getCurrentGitCommitSha());
+    return fs.writeFileSync('success');
+});
 gulp.task('deleteSuccessFile', function () {
     if(buildingFor.ios()){
         qmLog.info("Deleting success file messes up iOS build or so I'm told by my previous comments...");
@@ -1524,11 +1562,11 @@ function writeStaticDataFile(){
         ' else if(typeof qm !== "undefined"){qm.staticData = staticData;} else {module.exports = staticData;} ' +
         'if(typeof qm !== "undefined"){qm.stateNames = staticData.stateNames;}';
     try {
-        writeToFile('www/data/qmStaticData.js', string);
+        writeToFile(paths.www.staticData, string);
     } catch(e){
         qmLog.error(e.message + ".  Maybe www/data doesn't exist but it might be resolved when we copy from src");
     }
-    return writeToFile('src/data/qmStaticData.js', string);
+    return writeToFile(paths.src.staticData, string);
 }
 gulp.task('staticDataFile', ['getAppConfigs'], function () {
     return writeStaticDataFile();
@@ -2878,6 +2916,11 @@ gulp.task('uploadBuddyBuildToS3', ['zipBuild'], function () {
 // Need configureAppAfterNpmInstall or build-ios-app results in infinite loop
 gulp.task('configureAppAfterNpmInstall', [], function (callback) {
     qmLog.info('gulp configureAppAfterNpmInstall');
+    if(qmGulp.buildInfoHelper.alreadyMinified()){
+        qmLog.info("Not configuring app because already built this commit");
+        callback();
+        return;
+    }
     if(!buildingFor.web()){
         qmLog.info("Not configuring app after yarn install because we're building for mobile");
         callback();
@@ -2919,9 +2962,9 @@ gulp.task('configureApp', [], function (callback) {
         'copyIconsToWwwImg',
         'copyServiceWorkerAndLibraries',
         'setVersionNumberInFiles',
-        'createSuccessFile',
-        'verifyExistenceOfBuildInfo',
         'replaceRelativePathsWithAbsolutePaths',
+        'verifyExistenceOfBuildInfo',
+        'createSuccessFile',
         callback);
 });
 gulp.task('_chrome-in-src', ['getAppConfigs'], function (callback) {
